changed the project structure and renamed field
