+++ app/src/main/java/ru/apps65/learningproject/characters/presentation/CharacterScreen.kt
@@ -1,0 +1,5 @@
+package ru.apps65.learningproject.characters.presentation
+
+import com.github.terrakok.cicerone.androidx.FragmentScreen
+
+fun charactersScreen() = FragmentScreen { CharactersFragment() }
