+++ app/src/main/java/ru/apps65/learningproject/characters/presentation/CharactersFragment.kt
@@ -1,11 +1,11 @@
-package ru.apps65.learningproject.presentation.characterscreen
+package ru.apps65.learningproject.characters.presentation
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import ru.apps65.learningproject.databinding.FragmentCharactersBinding
-import ru.apps65.learningproject.presentation.common.BaseFragment
-import ru.apps65.learningproject.presentation.screens.Screens
+import ru.apps65.learningproject.common.BaseFragment
+import ru.apps65.learningproject.detailcharacter.presentation.characterDetailScreen
 import ru.apps65.learningproject.utils.getRouter
 
 class CharactersFragment: BaseFragment<FragmentCharactersBinding>() {
@@ -17,8 +17,8 @@ class CharactersFragment: BaseFragment<FragmentCharactersBinding>() {
     ): FragmentCharactersBinding =FragmentCharactersBinding.inflate(inflater, container, false)
 
     override fun otherSetups() {
-        binding.buttonToDetail.setOnClickListener {
-            getRouter().navigateTo(Screens.CharacterDetailScreen())
+        binding?.buttonToDetail?.setOnClickListener {
+            getRouter().navigateTo(characterDetailScreen())
         }
     }
 }
