+++ app/src/main/java/ru/apps65/learningproject/detailcharacter/presentation/CharacterDetailScreen.kt
@@ -1,0 +1,5 @@
+package ru.apps65.learningproject.detailcharacter.presentation
+
+import com.github.terrakok.cicerone.androidx.FragmentScreen
+
+fun characterDetailScreen() = FragmentScreen { CharacterDetailFragment() }
