+++ app/src/main/res/layout/activity_main.xml
@@ -5,5 +5,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
-    tools:context=".MainActivity"/>
+    tools:context=".main.MainActivity"/>
